Moved ClickGUI into client package
